propertyguru-crawler/pyproject.toml
[project]
name = "crawler-framework"
version = "1.0.0"
description = "安全高效的Python爬虫框架，支持代理IP、数据库和S3存储"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["crawler", "spider", "web-scraping", "proxy", "s3"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # HTTP请求和爬虫相关
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "selenium>=4.15.0",
    "undetected-chromedriver>=3.5.0",
    "setuptools>=65.0.0",
    "playwright>=1.40.0",

    # 代理和网络
    "PySocks>=1.7.1",

    # 数据库
    "pymongo>=4.6.0",
    "pymysql>=1.1.0",
    "psycopg2-binary>=2.9.9",  # PostgreSQL 驱动
    "sqlalchemy>=2.0.0",
    # "redis>=5.0.0",  # 可选：如果需要 Redis 支持，取消注释

    # AWS S3存储
    "boto3>=1.34.0",
    "botocore>=1.34.0",

    # 七牛云存储
    "qiniu>=7.11.0",

    # 异步支持
    "aiofiles>=23.2.0",

    # 虚拟显示（用于无窗口运行有头浏览器）
    "pyvirtualdisplay>=3.0",

    # 数据处理
    "pandas>=2.1.0",
    "numpy>=1.26.0",

    # 配置和环境
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",

    # 日志和监控
    "loguru>=0.7.2",

    # 工具库
    "tenacity>=8.2.3",
    "fake-useragent>=1.4.0",
    "python-dateutil>=2.8.2",

    # 加密和安全
    "cryptography>=41.0.0",

    # 并发和调度
    "schedule>=1.2.0",
    "apscheduler>=3.10.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.1.0",  # ruff 替代 flake8
    "mypy>=1.5.0",
    "isort>=5.12.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/crawler-framework"
Repository = "https://github.com/yourusername/crawler-framework"
Documentation = "https://github.com/yourusername/crawler-framework#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["crawler", "utils"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
    "pre-commit>=3.0.0",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "selenium.*",
    "boto3.*",
    "botocore.*",
    "playwright.*",
    "yaml.*",
    "requests.*",
    "lxml.*",
    "loguru.*",
    "bs4.*",
    "fake_useragent.*",
    "sqlalchemy.dialects.*",  # SQLAlchemy dialects 有动态属性
]
ignore_missing_imports = true
ignore_errors = true

[[tool.mypy.overrides]]
module = ["undetected_chromedriver", "undetected_chromedriver.*"]
ignore_missing_imports = true
ignore_errors = true

[[tool.mypy.overrides]]
module = ["redis", "redis.*"]
ignore_missing_imports = true
ignore_errors = true

[[tool.mypy.overrides]]
module = ["crawler.propertyguru_crawler"]
warn_unreachable = false

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports
"examples/*" = ["T201"]   # print statements in examples

[tool.ruff.lint.isort]
known-first-party = ["crawler", "utils"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
